Actually run tests synchronously
Replace the 'async' keyword with 'asynchronous' which is the actual name
of the Configuration property. There is no error for passing an invalid
option to Configuration.

Improve performance by running a shared server between test case methods
which is started with each class and shutdown after the case is
completed, clearing state in-between.