Move delivery handling into separate Delivery class
Instead of having notifications self-deliver, this introduces a Delivery
class which can deliver a payload and is provided an instance of
Configuration from which to derive its destination.

The two implementations of delivering a payload were moved into separate
classes, UrllibDelivery and RequestsDelivery, using urllib2/urllib3 or
requests respectively.

By default, if requests is present, it is used to deliver the payload,
otherwise it uses urllib