fix(handlers): Skip sending reports to Bugsnag from lib internals
In the case that Bugsnag is sending a log record, they are intended for
the user to see an internal issue or fix configuration problems.
In the specific case that the problem was a failure to send an error
report, there is a chance that this could cause a stack overflow if the
failure persists. This change avoids sending reports to Bugsnag from log
messages originating within the bugsnag library.