Fix for ignore_classes setting for django's 404
In my setup, by default, bugsnag logged some of Http404 despite the default config telling otherwise.

Looking at the examples, it was not super clear that I had to set the value to the full exception path in the setting. Then I managed to fix my setup by updating the exception path which I noticed was slightly different.

- Updating the examples to be clearer by specifying the full exception path
- Updating the default config to expect Http404 with full path or shortcut path