Remove dependency on distutils
distutils is vaguely "legacy" and on some systems (e.g. Ubuntu 18.04)
requires an additional system package (python3-distutils) to be
available, making it slightly harder to use bugsnag.

We can get the same kinds of information from sysconfig instead, which
is "more" a part of the stdlib and generally provides the same behavior,
e.g.:

    >>> distutils.sysconfig.get_python_lib()
    '/home/ckuehl/proj/buggo/venv/lib/python3.6/site-packages'
    >>> sysconfig.get_paths()['purelib']
    '/home/ckuehl/proj/buggo/venv/lib/python3.6/site-packages'

In the long term, I think it probably makes the most sense to actually
report all of `sys.path` along with errors, rather than just this one
value.