fix(utils): Make nested dict uniqueness checks work
References to nested dicts/field are recycled when going out of scope,
which is immediately since the references are not retained for the
lifetime of the sanitization pass. This change keeps references to seen
nested objects in addition to ignored objects to ensure id() returns
unique references.

Related reports on the Python issue tracker which explain the issue:
* https://bugs.python.org/issue36156
* https://bugs.python.org/issue33685

>>> help(id)

Help on built-in function id in module __builtin__:

id(...)
    id(object) -> integer

    Return the identity of an object.  This is guaranteed to be unique
    among simultaneously existing objects.  (Hint: it's the object's 
    memory address.)