fix(delivery): Match interface to existing implementations
Existing delivery impls use an options hash to send request-specific
parameters, however the delivery interface does not reflect this
requirement, which has effectively meant that any custom implementation
would be broken if implemented as-is. This change brings the parent
class in line with how its been used for the past three years.