Add 'skip_bugsnag_middleware' function
This is a callback that returns `False` when an event's `original_error`
field has a `skip_bugsnag` attr set to `True`. For example:

```python
exception = Exception('oh no')

bugsnag.notify(exception) # will notify

exception.skip_bugsnag = True

bugsnag.notify(exception) # will not notify
```

This is useful when you want to manually notify Bugsnag of an exception
but still re-raise it to be handled by another exception handler, e.g.
if there is useful metadata available only to the location where the
exception is raised