Relax checks on response status code
Currently we require a 200 status code for events and a 202 for sessions
but there's no need to be so specific as any 2xx status indicates
success

Now we only warn on codes outside of the 2xx range, unless a specific
code has been set in `options['success']`. If a specific code is given
then it must match exactly, i.e. it has the same behaviour as before